chore: resolve remaining review cleanups
Agent-Logs-Url: https://github.com/xinghao2003/side-project/sessions/f7fdd0ab-f4d7-4ee1-b224-53680ed42f48

+++ Capture Bot/include/Notifier.h
@@ -32,7 +32,7 @@ private:
   bool wifiReady_ = false;
   bool timeSynced_ = false;
   unsigned long lastTelegramCommandPollMs_ = 0;
-  uint32_t telegramUpdateOffset_ = 0;
+  unsigned long telegramUpdateOffset_ = 0;
   char retryPaths_[NetworkSettings::TelegramRetryQueueSize][48] = {};
   uint8_t retryAttempts_[NetworkSettings::TelegramRetryQueueSize] = {};
   unsigned long retryAfterMs_[NetworkSettings::TelegramRetryQueueSize] = {};
